fixed populate_form()
Not sure what the issue is here, but when you have a form open and click on another date it ignores the return and tries to run the function (-> crash). I think the issue is that the selection change is detected but working on the previous list, so adding another list_populate amends that. Then it cannot pickup empty lists with a -1 return as suggested by the author, changed that to the count() method.